% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrs.R
\name{ntrs_citation}
\alias{ntrs_citation}
\title{Fetch a single NTRS citation by ID}
\usage{
ntrs_citation(id)
}
\arguments{
\item{id}{NTRS citation ID}
}
\value{
tibble: single row with citation details
}
\description{
Fetch a single NTRS citation by ID
}
